Correct certs and DNS records
