+++ cloud/bin/cloud.ts
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-import { App, Environment } from 'aws-cdk-lib';
+import { App, Environment } from 'aws-cdk-lib/core';
 import 'source-map-support/register';
 
 import {
@@ -9,11 +9,16 @@ import {
 	stackName,
 	ApiStack,
 	AuthStack,
-	RoutingStack,
+	CertificateStack,
+	HostedZoneStack,
 	UiStack,
 } from '../lib';
 
 const app = new App();
+const generateStackName = stackName(app);
+const generateDescription = resourceDescription(app);
+
+/* Common stack resources */
 
 const env: Environment = {
 	account: process.env.CDK_DEFAULT_ACCOUNT,
@@ -27,21 +32,35 @@ const tags = {
 	'keep-alive': '9-5-without-weekends',
 };
 
-const generateStackName = stackName(app);
-const generateDescription = resourceDescription(app);
+/* Stack constructs */
 
-const routingStack = new RoutingStack(app, generateStackName('routing'), {
-	description: generateDescription('Route 53 stack'),
-	env,
-	tags,
-});
+const hostedZoneStack = new HostedZoneStack(
+	app,
+	generateStackName('hosted-zone'),
+	{
+		description: generateDescription('Hosted Zone stack'),
+		env,
+		tags,
+	}
+);
+
+const certificateStack = new CertificateStack(
+	app,
+	generateStackName('certificate'),
+	{
+		description: generateDescription('Certificate stack'),
+		env,
+		tags,
+		hostedZone: hostedZoneStack.hostedZone,
+	}
+);
 
 const uiStack = new UiStack(app, generateStackName('ui'), {
 	description: generateDescription('UI stack'),
 	env,
 	tags,
-	certificate: routingStack.certificate,
-	hostedZone: routingStack.hostedZone,
+	certificate: certificateStack.cloudFrontCert,
+	hostedZone: hostedZoneStack.hostedZone,
 });
 
 /*const authStack = */ new AuthStack(app, generateStackName('auth'), {
@@ -55,8 +74,8 @@ new ApiStack(app, generateStackName('api'), {
 	description: generateDescription('API stack'),
 	env,
 	tags,
-	certificate: routingStack.certificate,
-	hostedZone: routingStack.hostedZone,
+	certificate: certificateStack.loadBalancerCert,
+	hostedZone: hostedZoneStack.hostedZone,
 	// userPool: authStack.userPool,
 	// userPoolClient: authStack.userPoolClient,
 	// userPoolDomain: authStack.userPoolDomain,
