+++ cloud/cdk.context.json
@@ -3,5 +3,6 @@
 		"eu-north-1a",
 		"eu-north-1b",
 		"eu-north-1c"
-	]
+	],
+	"ami:account=992382568770:filters.image-type.0=machine:filters.name.0=amzn-ami-vpc-nat-*:filters.state.0=available:owners.0=amazon:region=eu-north-1": "ami-072517490bf2cf3a3"
 }
