+++ cloud/lib/index.ts
@@ -1,5 +1,7 @@
 export * from './resourceNamingUtils';
 export { ApiStack } from './api-stack';
 export { AuthStack } from './auth-stack';
-export { RoutingStack } from './routing-stack';
+export { CertificateStack } from './certificate-stack';
+export { HostedZoneStack } from './hostedzone-stack';
+export { SchedulerStack } from './scheduler-stack';
 export { UiStack } from './ui-stack';
